+++ src/multiband_compressor.cpp
@@ -23,11 +23,49 @@ MultibandCompressor::MultibandCompressor(const std::string& tag,
                                          const std::string& schema,
                                          const std::string& schema_path,
                                          PipeManager* pipe_manager)
-    : PluginBase(tag, plugin_name::multiband_compressor, schema, schema_path, pipe_manager),
-      lv2_wrapper(std::make_unique<lv2::Lv2Wrapper>("http://lsp-plug.in/plugins/lv2/mb_compressor_stereo")) {
+    : PluginBase(tag, plugin_name::multiband_compressor, schema, schema_path, pipe_manager, true),
+      lv2_wrapper(std::make_unique<lv2::Lv2Wrapper>("http://lsp-plug.in/plugins/lv2/sc_mb_compressor_stereo")) {
   if (!lv2_wrapper->found_plugin) {
-    util::debug(log_tag + "http://lsp-plug.in/plugins/lv2/mb_compressor_stereo is not installed");
+    util::debug(log_tag + "http://lsp-plug.in/plugins/lv2/sc_mb_compressor_stereo is not installed");
   }
+
+  auto update_sidechain_links = [=, this](const auto& key) {
+    auto external_sidechain_enabled = false;
+
+    for (uint n = 0U; !external_sidechain_enabled && n < n_bands; n++) {
+      const auto nstr = std::to_string(n);
+
+      external_sidechain_enabled = settings->get_boolean("external-sidechain" + nstr);
+    }
+
+    if (external_sidechain_enabled) {
+      const auto device_name = settings->get_string("sidechain-input-device").raw();
+
+      NodeInfo input_device = pm->ee_source_node;
+
+      for (const auto& [ts, node] : pm->node_map) {
+        if (node.name == device_name) {
+          input_device = node;
+
+          break;
+        }
+      }
+
+      pm->destroy_links(list_proxies);
+
+      list_proxies.clear();
+
+      for (const auto& link : pm->link_nodes(input_device.id, get_node_id(), true)) {
+        list_proxies.push_back(link);
+      }
+    } else {
+      pm->destroy_links(list_proxies);
+
+      list_proxies.clear();
+    }
+  };
+
+  settings->signal_changed("sidechain-input-device").connect(update_sidechain_links);
 
   lv2_wrapper->bind_key_enum(settings, "compressor-mode", "mode");
 
@@ -35,6 +73,10 @@ MultibandCompressor::MultibandCompressor(const std::string& tag,
 
   for (uint n = 0U; n < n_bands; n++) {
     const auto nstr = std::to_string(n);
+
+    settings->signal_changed("external-sidechain" + nstr).connect(update_sidechain_links);
+
+    lv2_wrapper->bind_key_bool(settings, "external-sidechain" + nstr, "sce_" + nstr);
 
     if (n > 0U) {
       lv2_wrapper->bind_key_bool(settings, "enable-band" + nstr, "cbe_" + nstr);
@@ -110,7 +152,9 @@ void MultibandCompressor::setup() {
 void MultibandCompressor::process(std::span<float>& left_in,
                                   std::span<float>& right_in,
                                   std::span<float>& left_out,
-                                  std::span<float>& right_out) {
+                                  std::span<float>& right_out,
+                                  std::span<float>& probe_left,
+                                  std::span<float>& probe_right) {
   if (!lv2_wrapper->found_plugin || !lv2_wrapper->has_instance() || bypass) {
     std::copy(left_in.begin(), left_in.end(), left_out.begin());
     std::copy(right_in.begin(), right_in.end(), right_out.begin());
@@ -122,7 +166,7 @@ void MultibandCompressor::process(std::span<float>& left_in,
     apply_gain(left_in, right_in, input_gain);
   }
 
-  lv2_wrapper->connect_data_ports(left_in, right_in, left_out, right_out);
+  lv2_wrapper->connect_data_ports(left_in, right_in, left_out, right_out, probe_left, probe_right);
   lv2_wrapper->run();
 
   if (output_gain != 1.0F) {